requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4
Click==7.0
coverage==4.5.1
Flask==1.0.2
Flask-Cors==3.0.6
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-Testing==0.7.1
gunicorn==19.9.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
nose==1.3.7
pathlib2==2.3.2
pluggy==0.8.0
psycopg2-binary==2.7.5
py==1.7.0
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.9.1
pytest-cov==2.6.0
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
validate-email==1.3
Werkzeug==0.14.1
wrapt==1.10.11
